% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eset-utils.R
\name{assayDataMelt}
\alias{assayDataMelt}
\title{Melting Assay Data}
\usage{
assayDataMelt(object, elements = NULL, varnames = c("feature_id", "sample_id"))
}
\arguments{
\item{object}{an object with a suitable \link[Biobase:assayDataElement]{Biobase::assayDataElement} method,
typically an \link[Biobase:ExpressionSet-class]{Biobase::ExpressionSet} object}

\item{elements}{an optional character vector that indicates the names of the
assay data elements to include in the result}

\item{varnames}{a character vector with 2 unique elements that indicates the name of the columns
that will hold the feature and sample names respectively.}
}
\value{
a \code{data.frame} object with the following structure:
\itemize{
\item character columns \code{feature_id} and \code{sample_id} containing the feature and sample names respectively;
\item other columns are numeric and correspond to each assay data element in the input object.
}
}
\description{
Melting Assay Data
}
