% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{bioc-generic}
\alias{bioc-generic}
\alias{featureNames,matrix-method}
\alias{featureNames<-,matrix-method}
\alias{sampleNames<-,matrix,ANY-method}
\alias{sampleNames,matrix,ANY}
\alias{sampleNames,matrix-method}
\alias{exprs,matrix-method}
\alias{exprs<-,matrix,ANY-method}
\title{Extracting Feature Names}
\usage{
\S4method{featureNames}{matrix}(object)

\S4method{featureNames}{matrix}(object) <- value

\S4method{sampleNames}{matrix,ANY}(object) <- value

\S4method{sampleNames}{matrix}(object)

\S4method{exprs}{matrix}(object)

\S4method{exprs}{matrix,ANY}(object) <- value
}
\description{
The \pkg{xbioc} package provides extra methods for the generic
\code{\link[Biobase]{featureNames}} and \code{\link[Biobase]{sampleNames}},
that extend the original Bioconductor interface.
}
\section{Functions}{
\itemize{
\item \code{featureNames,matrix-method}: featureNames Returns the row names of \code{object}.

\item \code{featureNames<-,matrix-method}: Set the row names of \code{object}.

\item \code{sampleNames<-,matrix,ANY-method}: Set the row names of \code{object}.

\item \code{sampleNames,matrix-method}: Returns the column names of \code{object}.

\item \code{exprs,matrix-method}: Simply returns \code{object}.
This method is defined so the generic can be called on both matrix and
\code{ExpressionSet} objects.

\item \code{exprs<-,matrix,ANY-method}: Simply assigns \code{value} to \code{object}.
This method is defined so the generic can be called on both matrix and
\code{ExpressionSet} objects.
}}

