% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{featureNames}
\alias{featureNames}
\alias{featureNames<-}
\alias{sampleNames}
\alias{sampleNames<-}
\alias{exprs}
\alias{exprs,matrix-method}
\title{Extracting Feature Names}
\usage{
featureNames(object)

featureNames(object) <- value

sampleNames(object)

sampleNames(object) <- value

exprs(object)

\S4method{exprs}{matrix}(object)
}
\arguments{
\item{object}{Object, possibly derived from class \code{eSet}.}

\item{value}{Character vector containing feature or sample names.}
}
\description{
The \pkg{xbioc} package provides extra methods for the generic
\code{\link[Biobase]{featureNames}} and \code{\link[Biobase]{sampleNames}},
that extend the original Bioconductor interface.

Simply returns \code{object}.
This method is defined so that function can seamlessly handle both matrix and
\code{ExpressionSet} objects.
}
