% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{bioc-utils}
\alias{bioc-utils}
\alias{isExpressionSet}
\alias{isMatrixData}
\title{Generic Bioconductor Utilities}
\usage{
isExpressionSet(x)

isMatrixData(x)
}
\arguments{
\item{x}{object}
}
\description{
Utilities to test Bioconductor object classes.
}
\section{Functions}{
\itemize{
\item \code{isExpressionSet}: tells if an object inherits from the class \link[Biobase:ExpressionSet]{Biobase::ExpressionSet}.

\item \code{isMatrixData}: tells if an object is a matrix-like data.
}}

