% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{cbind_ExpressionSet}
\alias{cbind_ExpressionSet}
\alias{rbind_ExpressionSet}
\title{Combining ExpressionSet Objects}
\usage{
cbind_ExpressionSet(..., .id = ".id", deparse.level = 1)

rbind_ExpressionSet(..., .id = "dataType", prefix_size = 0L)
}
\description{
\code{cbind_ExpressionSet} is a direct shortcut to \link{cbind.ExpressionSet}.
See argument details therein.
}
