% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{checkSYMBOLS}
\alias{checkSYMBOLS}
\title{Checks Gene Symbol Validity}
\usage{
checkSYMBOLS(x, value = FALSE, replace = FALSE, quiet = TRUE)
}
\arguments{
\item{x}{Character vector of gene symbols to check}

\item{value}{Logical that indicates if the value of invalid symbols
should be returned (\code{TRUE}), or only their index (\code{FALSE}).}

\item{replace}{logical that indicates if fixable symbols should be replaced
with their correct value.
If \code{value = FALSE}, then the corrected symbols are set as names of the
result integer vector, and the incorrect (input) symbols as values.
If \code{value = FALSE}, then the corrected symbols are returned as values of the
result character vector, and the incorrect (input) symbols as names.}
}
\value{
character or integer vector depending on the value of argument
\var{value}.
See help for argument \var{replace}, as this affects what goes into
names and values of the result vector.
}
\description{
Look for gene symbols that were transformed by Excel nasty default
setting, which interpreted them as dates or numeric.
}
\details{
This issue, although well known in the bioinformatics community, regularly
surfaces to the detriment of the person who analyses the data.
It has been reported to affect several publications and even large genomic projects
(Zeeberg2004, Ziemann et al. (2016)).
}
\examples{
\dontshow{
options(R_CHECK_RUNNING_EXAMPLES_=TRUE) ## roxygen generated flag
}


bad_symbols <- c('AAA', 'BBB', '1-Dec', '10-JAN')
# index of incorrect symbols
checkSYMBOLS(bad_symbols)
# incorrect symbols
checkSYMBOLS(bad_symbols, value = TRUE)
# return fixed symbols as names (index vector)
checkSYMBOLS(bad_symbols, replace = TRUE)
# return fixed symbols as values (mapping)
checkSYMBOLS(bad_symbols, value = TRUE, replace = TRUE)

}
\references{
Zeeberg2004

Ziemann M, Eren Y and El-osta A (2016). “Gene name errors are widespread in the scientific literature.” \emph{Genome Biology}, pp.
17-19. ISSN 1474-760X, doi: 10.1186/s13059-016-1044-7 (URL: http://doi.org/10.1186/s13059-016-1044-7), <URL:
http://dx.doi.org/10.1186/s13059-016-1044-7>.
}
