% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_append}
\alias{df_append}
\title{Appending Columns to Data Frames}
\usage{
df_append(x, ..., after = length(x), names = NULL,
  stringsAsFactors = getOption("stringsAsFactors", FALSE))
}
\arguments{
\item{x}{a \code{data.frame} object}

\item{...}{variables to be appended to \var{x}.}

\item{after}{index or column name after which the new variables should be inserted.
Default is to append the columns at the end of the \code{data.frame}.}

\item{names}{optional names for the newly inserted columns.}

\item{stringsAsFactors}{logical that indicates if the newly inserted string variables should be converted to
factors.}
}
\value{
a \code{data.frame}
}
\description{
This function performs on \code{data.frame} objects a similar operation as
\code{\link[base]{append}} does on \code{list} objects.
That is it enables prepending/inserting/appending columns at specific positions.
}
\examples{
\dontshow{
options(R_CHECK_RUNNING_EXAMPLES_=TRUE) ## roxygen generated flag
}


x <- data.frame(a = 1:4, b = letters[1:4])
# append at the end
df_append(x, c = runif(4))

# append at the beginning
df_append(x, c = runif(4), after = 0L)

# append after first column
df_append(x, c = runif(4), after = 1L)
df_append(x, c = runif(4), after = 'a')

}
