% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{digestN}
\alias{digestN}
\title{Computing a Short Digest String}
\usage{
digestN(x, n = 7, ...)
}
\arguments{
\item{x}{an object}

\item{n}{a single integer/numeric that indicates the desired length of the digest string.
Use \code{n = Inf} to obtain full digest strings.}

\item{...}{other arguments passed to \link[digest:digest]{digest::digest}}
}
\value{
a character vector of digest strings.
}
\description{
Computing a Short Digest String
}
\examples{

x <- list(1, 2, 3)
digestN(x)
# longer diest
digestN(x, 15)
# full digest
digestN(x, Inf)

}
