% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marray.R
\name{featureIsControl}
\alias{featureIsControl}
\title{Extracting Control Probes}
\usage{
featureIsControl(object)
}
\arguments{
\item{object}{expression data as a \code{matrix} or \code{ExpressionSet} object.}
}
\description{
Extracting Control Probes
}
