% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{flow_reference<-}
\alias{flow_reference<-}
\title{Change Reference Populations in Flowcytometry Data}
\usage{
flow_reference(x, ref, parent, ...) <- value
}
\arguments{
\item{x}{flow data}

\item{ref}{indicates the reference population, that is either \code{'parent'} or
\code{'total'}}
}
\description{
Change Reference Populations in Flowcytometry Data
}
