% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{intersect}
\alias{intersect}
\alias{intersect,MatrixData,logical-method}
\alias{intersect,MatrixData,character-method}
\alias{intersect,MatrixData,MatrixData-method}
\title{Enhanced Subsetting for Matrix-like Data}
\usage{
\S4method{intersect}{MatrixData,logical}(x, y)

\S4method{intersect}{MatrixData,character}(x, y)

\S4method{intersect}{MatrixData,MatrixData}(x, y)
}
\arguments{
\item{x}{vectors (of the same mode) containing a sequence
    of items (conceptually) with no duplicated values.}

\item{y}{vectors (of the same mode) containing a sequence
    of items (conceptually) with no duplicated values.}
}
\description{
These methods subset Matrix-like data only keeping rows
whose names are in common with some other data.

This method is equivalent to \code{\link[base]{subset}(x, y)}.

Subset a matrix-like object by only keeping the rows whose
names are in a given reference character vector.

This is a shortcut for \code{intersect(x, featureNames(y), ...)}.
}
