% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_logscale}
\alias{is_logscale}
\alias{has_logscale_outliers}
\alias{log_transform}
\title{Detect Log-transformed Data}
\source{
\url{www.ncbi.org/geo}
}
\usage{
is_logscale(x, robust = TRUE, extremes = c(0.02, 0.98))

has_logscale_outliers(x, ...)

log_transform(x, base = 2)
}
\arguments{
\item{x}{a numeric data object (matrix, vector, ExpressionSet)}

\item{robust}{logical that indicates if the decision should be
done robustly by removing the more extreme values.}

\item{extremes}{a length-2 numeric vector that gives the lower and upper
quantiles used to remove extreme values.
Only relevant when \code{robust = TRUE}.}

\item{...}{other arguments passed to \link{is_logscale}, like argument \code{extremes}
to control the extreme values that are removed.}

\item{base}{log base to use.}
}
\description{
\code{is_logscale} tells if some numeric data is in log scale,
e.g., normalized microarray data, using the same heuristic as GEO2R.

\code{log_transform} apply a log transformation to the data.
Negative values get assigned the value \code{\link{NaN}}.
}
\details{
The data needs to be of reasonable size and variance for the detection
heuristic to work correctly.
}
\section{Functions}{
\itemize{
\item \code{has_logscale_outliers}: tests if there are outliers to the log-scale distribution.
}}

\examples{
\dontshow{
options(R_CHECK_RUNNING_EXAMPLES_=TRUE) ## roxygen generated flag
}


x <- matrix(rnorm(20*10, mean=500), 20, 10)
is_logscale(x)
is_logscale(log_transform(x))

}
