% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{isExpressionSet}
\alias{isExpressionSet}
\alias{isMatrixData}
\title{Generic Bioconductor Utilities}
\usage{
isExpressionSet(x)

isMatrixData(x)
}
\arguments{
\item{x}{object}
}
\description{
\code{isExpressionSet} tells if an object inherits from the class \code{\link[Biobase]{ExpressionSet}}.

\code{isExpressionSet} tells if an object is a matrix-like data.
}
