% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eset-utils.R
\name{melt.ExpressionSet}
\alias{melt.ExpressionSet}
\title{Melt an ExpressionSet for Use with ggplot}
\usage{
melt.ExpressionSet(data, ..., pData = pData(data), na.rm = FALSE,
  value.name = "value", extra = list())
}
\arguments{
\item{data}{an \code{ExpressionSet} object.}

\item{...}{other data types to be melt.
Named arguments, will result in separate variables in the returned \code{data.frame}.}

\item{pData}{phenotypic data to attach to the result \code{data.frame}.}

\item{na.rm}{Should NA values be removed from the data set? This will
convert explicit missings to implicit missings.}

\item{value.name}{name of variable used to store values}

\item{extra}{list of other matrix-like data to melt and \code{rbind.fill} to the melted primary data.}
}
\description{
Melt the expression data of an \code{\link[Biobase]{ExpressionSet}},
also binding any corresponding phenotypic data.
}
