% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NAmap}
\alias{NAmap}
\title{Initialising a Mapping List}
\usage{
NAmap(x)
}
\arguments{
\item{x}{object used as right keys}
}
\value{
a list with names \code{x} mapping to NA values.
}
\description{
Initialising a Mapping List
}
\examples{
\dontshow{
options(R_CHECK_RUNNING_EXAMPLES_=TRUE) ## roxygen generated flag
}

NAmap(letters[1:5])

}
