% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eset-utils.R
\name{pVar}
\alias{pVar}
\alias{pVar,matrix-method}
\alias{pVar,ExpressionSet-method}
\title{Retrieve Phenotypic Variable from Data}
\usage{
pVar(object, x, ...)

\S4method{pVar}{matrix}(object, x)

\S4method{pVar}{ExpressionSet}(object, x)
}
\arguments{
\item{object}{an \code{ExpressionSet} object.}

\item{x}{variable as an atomic vector or as a string corresponding to the
name of a phenotypic variable to be looked up in \code{phenData(object)}.}
}
\description{
This is a generic function to extract a phenotypic variable from data objects.
It aims is to encompass variables that are directly defined as vectors or,
for example, specified as a name that is looked up in the \code{phenoData} variables
of an \code{ExpressionSet} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: simply returns \code{x}, after checking that it has the same length
as the number of columns in \code{x}.

\item \code{ExpressionSet}: either return \code{x} or the phenotypic variable \code{pData(object)[[x]]}
if it exists.
}}

