% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{read.fcs}
\alias{read.fcs}
\title{Read All FCS Files in a Directory}
\usage{
read.fcs(path, ..., pattern = "\\\\.fcs$", recursive = FALSE)
}
\arguments{
\item{path}{path to the directory to load}

\item{...}{arguments passed to \code{\link[flowCore]{read.FCS}}}

\item{pattern}{file pattern to match.
Files are matched using regular expressions via
\code{\link{list.files}} in case insensitive mode.}

\item{recursive}{logical that indicates if sub-directory should
be looked up.}
}
\description{
Read all the FCS files in a directory, using
\code{\link[flowCore]{read.FCS}}.
}
