% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{samFDR}
\alias{samFDR}
\title{Compute Permutation-based FDR as in SAM}
\usage{
samFDR(obs, perms, alternative = c("two.sided", "greater", "less"), digit = 2)
}
\arguments{
\item{obs}{observed statistics}

\item{perms}{matrix of statistics obtained for each permutation (column).}

\item{alternative}{alternative hypothesis to test}

\item{digit}{FDR precision in number of digits}
}
\description{
Compute Permutation-based FDR as in SAM
}
