\docType{package}
\name{xbioc-package}
\alias{xbioc-package}
\title{Extra Base Functions for Bioconductor}
\description{
  Provides extra utility functions to perform common tasks
  in the analysis of omic data, leveraging and enhancing
  features provided by Bioconductor packages.
}
\details{
  \tabular{ll}{ Package: \tab xbioc\cr Type: \tab
  Package\cr Version: \tab 0.1.4\cr Date: \tab
  2014-07-03\cr License: \tab GPL (>= 3)\cr LazyLoad: \tab
  yes\cr }
}
\author{
  Renaud Gaujoux <renaud@tx.technion.ac.il> (Maintainer)
}

