% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{annotation-utils}
\alias{annotation-utils}
\alias{hasAnnotation}
\alias{getAnnotation}
\alias{getAnnotation.default}
\alias{setAnnotation}
\alias{setAnnotation.default}
\alias{getAnnotation.list}
\alias{setAnnotation.list}
\title{Handling Object Annotations}
\usage{
hasAnnotation(object, ...)

getAnnotation(object, ...)

\method{getAnnotation}{default}(object, ..., null = TRUE)

setAnnotation(object, value, ...)

\method{setAnnotation}{default}(object, value, as.attribute = TRUE, ...)

\method{getAnnotation}{list}(object, ...)

\method{setAnnotation}{list}(object, value, force = FALSE, ...)
}
\arguments{
\item{object}{an object}

\item{...}{extra parameters (currently not used)}

\item{null}{logical that indicates if an empty character string should
be return as \code{NULL}.}

\item{value}{new annotation string, e.g., \code{'hgu133plu2.db'}.}

\item{as.attribute}{logical that indicates that the annotation string can be
stored as an attribute, if no suitable \code{`annotation <-`} method is found.}

\item{force}{single logical that indicates if the annotation string should be set
even if the objects do not share the same original annotation.}
}
\description{
Annotation utility functions.
}
\section{Functions}{
\itemize{
\item \code{hasAnnotation}: tells if an object has some -- non empty -- attached annotation.

\item \code{getAnnotation}: try extracting embedded annotations.

\item \code{getAnnotation.default}: the default method returns \code{NULL} if the object contains
no annotation.

\item \code{setAnnotation}: embbeds an annotation string into an object
(typically the name of an annotation package, e.g., \code{'hgu133plus2.db'}).

\code{setAnnotation} uses a suitable \code{annotation<-} method if it
exists, or sets the annotation string into attribute \code{'annotation'}.
It returns the modified object.

\item \code{getAnnotation.list}: gets annotation for each element in a list of \code{ExpressionSet}
objects.

\item \code{setAnnotation.list}: set annotation on each element of a list of \code{ExpressionSet} objects.
}}

\section{Methods (by generic)}{
\bold{getAnnotation:}

\itemize{
\item \code{getAnnotation(default)}: the default method returns \code{NULL} if the object contains
no annotation.

\item \code{getAnnotation(list)}: gets annotation for each element in a list of \code{ExpressionSet}
objects.
}


\bold{setAnnotation:}

\itemize{
\item \code{setAnnotation(list)}: set annotation on each element of a list of \code{ExpressionSet} objects.
}

}

