% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-packages.R
\name{biocann_orgdb}
\alias{biocann_orgdb}
\alias{biocann_orgdb_pkgname}
\title{Bioconductor Organism Data and Packages}
\usage{
biocann_orgdb(organism, optional = FALSE)

biocann_orgdb_pkgname(organism)
}
\arguments{
\item{organism}{Organism name (case insensitive).
Can be specified using the common name (e.g., mouse), a partial match of the latin name (e.g., Mus musculus),
or its abbreviation (e.g., Mm).
It can also be a \code{AnnDbBimap} object like \code{org.Hs.egSYMBOL2EG}.

The following organisms are currently supported:

human, arabidopsis, bovine, canine, chicken, chimp, malaria, mouse, pig, rat, rhesus, worm, zebrafish}

\item{optional}{logical that indicates if the function should raise an error if no package is found
for the given organism specification (\code{optional=FALSE}), or return \code{NA} data and throw a warning.}
}
\description{
Bioconductor Organism Data and Packages
}
\section{Functions}{
\itemize{
\item \code{biocann_orgdb_pkgname}: Returns the organism annotation package for a given
organism.
}}

\examples{
\dontshow{
options(R_CHECK_RUNNING_EXAMPLES_=TRUE) ## roxygen generated flag
}


# list all 
biocann_orgdb()

# common name
biocann_orgdb('human')

# latin name
biocann_orgdb('canis')
# any partial match works
biocann_orgdb('canis lupus')

# abbreviation
biocann_orgdb('Mm')

}
