% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\docType{methods}
\name{compareFeatures}
\alias{compareFeatures}
\alias{compareFeatures,character,character-method}
\title{Compare Features Sets}
\usage{
compareFeatures(x, y, ...)

\S4method{compareFeatures}{character,character}(x, y, unique = FALSE)
}
\arguments{
\item{x, y}{objects with associated featureNames}

\item{...}{extra arguments eventually passed to \code{compareFeatures,character,character}.}
}
\description{
Compares the set of features associated with an object, e.g., \code{\link[Biobase]{ExpressionSet}}
or \code{\link[AnnotationDbi]{ChipDb-class}} objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = character,y = character}: Workhorse function comparing the feature sets \code{x} and \code{y}.


}}

