% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{convertAlias}
\alias{convertAlias}
\title{Converts Gene Alias to Official Gene Symbols}
\usage{
convertAlias(x, nomatch = NA, organism = "Hs", verbose = TRUE)
}
\arguments{
\item{x}{character vector of gene aliases}

\item{nomatch}{value to use for those aliases that could not
be mapped.
This argument can also be set to \code{TRUE} to make unmapped aliases
keep their original input values.}

\item{organism}{name of the reference organism.
Currently only Human ('Hs') is supported.}

\item{verbose}{logical that indicates if mapped and unmapped symbols
should be output to the \code{stderr}.}
}
\value{
a character vector of the mapped symbols, whose names correspond to
the original values in \code{x}.
}
\description{
Converts Gene Alias to Official Gene Symbols
}
\examples{
\dontshow{
options(R_CHECK_RUNNING_EXAMPLES_=TRUE) ## roxygen generated flag
}


convertAlias(c('A', 'IL10'))
convertAlias(c('A', 'IL-10', 'IL1RA'))
convertAlias(c('A', 'IL-10', 'IL1RA'), nomatch = TRUE)

}
