% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\docType{methods}
\name{log,ExpressionSet-method}
\alias{log,ExpressionSet-method}
\alias{expb}
\alias{exp,ExpressionSet-method}
\alias{range,ExpressionSet-method}
\alias{quantile.ExpressionSet}
\title{Numeric Computations on ExpressionSet objects}
\usage{
\S4method{log}{ExpressionSet}(x, ...)

expb(x, ...)

\S4method{exp}{ExpressionSet}(x)

\S4method{range}{ExpressionSet}(x, ..., na.rm = FALSE)

\method{quantile}{ExpressionSet}(x, ...)
}
\arguments{
\item{x}{an \code{ExpressionSet} object.}

\item{...}{extra arguments passed to subsequent calls, usually of the corresponding
method in the \pkg{stats} package.}
}
\description{
The \pkg{xbioc} package defines some generics and methods to apply numeric transformations
to \code{ExpressionSet} objects, which is convenient when working on gene expression
deconvolution algorithms, where scale (log/linear) may matter.

\code{log} log-transforms the expression matrix of \code{\link{ExpressionSet}} objects.

\code{expb} applies an entry-wise exponential transformation to the expression matrix
of \code{\link{ExpressionSet}}, in a specific base.

\code{exp} is equivalent to \code{expb(x, exp(1))}.

\code{range} computes the range of expression values from
\code{\link{ExpressionSet}} objects.

\code{quantile} computes the range of expression values in
\code{\link{ExpressionSet}} objects.
}
