\name{isExpressionSet}
\alias{isExpressionSet}
\alias{isMatrixData}
\title{Generic Bioconductor Utilities}
\usage{
  isExpressionSet(x)

  isMatrixData(x)
}
\arguments{
  \item{x}{object}
}
\description{
  \code{isExpressionSet} tells if an object inherits from
  the class \code{\link[Biobase]{ExpressionSet}}.

  \code{isExpressionSet} tells if an object is a
  matrix-like data.
}

